# frontend/packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NookOS API",
    "description": "The open workspace operating system for modern software teams.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/nodes/{id}": {
      "get": {
        "tags": [
          "crate::routes::nodes"
        ],
        "operationId": "get_node",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "crate::routes::nodes"
        ],
        "operationId": "delete_node",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/tasks/{id}/dispatch": {
      "post": {
        "tags": [
          "crate::routes::taskwork"
        ],
        "operationId": "task_dispatch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/move": {
      "post": {
        "tags": [
          "crate::routes::taskwork"
        ],
        "operationId": "task_move",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/prune-worktree": {
      "post": {
        "tags": [
          "crate::routes::taskwork"
        ],
        "operationId": "task_prune_worktree",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/start-work": {
      "post": {
        "tags": [
          "crate::routes::taskwork"
        ],
        "operationId": "task_start_work",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartWorkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartWorkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/submit-pr": {
      "post": {
        "tags": [
          "crate::routes::taskwork"
        ],
        "operationId": "task_submit_pr",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPrRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces": {
      "get": {
        "tags": [
          "crate::routes::workspaces"
        ],
        "operationId": "list_workspaces",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceDetail"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::workspaces"
        ],
        "operationId": "create_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{id}": {
      "get": {
        "tags": [
          "crate::routes::workspaces"
        ],
        "operationId": "get_workspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDetail"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/workspaces/{id}/git": {
      "get": {
        "tags": [
          "crate::routes::workspaces"
        ],
        "summary": "Live git status + working-tree diff, relayed from the node.",
        "operationId": "workspace_git_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NodeId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ColumnId": {
        "type": "string",
        "format": "uuid"
      },
...
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "GitFileStatus": {
        "type": "object",
        "required": [
          "status",
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Porcelain status code, e.g. \" M\", \"??\", \"A \"."
          }
        }
      },
      "GitStatusResponse": {
        "type": "object",
        "required": [
          "dirty",
          "files",
          "diff"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "diff": {
            "type": "string",
            "description": "Unified diff of the working tree (truncated by the node if huge)."
          },
          "dirty": {
            "type": "boolean"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GitFileStatus"
            }
          }
        }
      },
...
      "MoveTaskRequest": {
        "type": "object",
        "required": [
          "column"
        ],
        "properties": {
          "column": {
            "type": "string"
          }
        }
      },
      "Node": {
        "type": "object",
        "description": "Status values: `online` | `offline`.",
        "required": [
          "id",
          "tenant_id",
          "name",
          "hostname",
          "platform",
          "capabilities",
          "resources",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "capabilities": {},
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "hostname": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/NodeId"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "resources": {
            "description": "Latest heartbeat resource sample (see `NodeResources`); `{}` until first."
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "Session": {
        "type": "object",
        "description": "Status values: `starting` | `running` | `detached` | `exited` | `error`.\nRuntime is an open string: \"claude\", \"hermes\", \"codex\", \"bash\", \"zsh\", ...",
        "required": [
          "id",
          "tenant_id",
          "workspace_id",
          "node_id",
          "name",
          "runtime",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserId"
              }
            ]
          },
          "ended_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/SessionId"
          },
          "name": {
            "type": "string"
          },
          "node_id": {
            "$ref": "#/components/schemas/NodeId"
          },
          "runtime": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          },
          "tmux_session": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "workspace_id": {
            "$ref": "#/components/schemas/WorkspaceId"
          }
        }
      },
      "SessionId": {
        "type": "string",
        "format": "uuid"
      },
...
      "StartWorkRequest": {
        "type": "object",
        "required": [
          "runtime"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "node_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NodeId"
              }
            ]
          },
          "runtime": {
            "type": "string"
          },
          "workspace_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorkspaceId"
              }
            ]
          }
        }
      },
      "StartWorkResponse": {
        "type": "object",
        "required": [
          "task",
          "session"
        ],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/Session"
          },
          "task": {
            "$ref": "#/components/schemas/TaskItem"
          }
        }
      },
      "SubmitPrRequest": {
        "type": "object",
        "properties": {
          "pr_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TaskId": {
        "type": "string",
        "format": "uuid"
      },
      "TaskItem": {
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "board_id",
          "column_id",
          "title",
          "position",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "assigned_node_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NodeId",
                "description": "Node the triage scheduler chose (or you forced) to run this work."
              }
            ]
          },
          "assignee_user_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserId"
              }
            ]
          },
          "board_id": {
            "$ref": "#/components/schemas/BoardId"
          },
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "column_id": {
            "$ref": "#/components/schemas/ColumnId"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "external_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "external_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/TaskId"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "pr_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SessionId"
              }
            ]
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "workspace_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorkspaceId"
              }
            ]
          },
          "worktree_node_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NodeId"
              }
            ]
          },
          "worktree_path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TenantId": {
        "type": "string",
        "format": "uuid"
      },
...
      "UserId": {
        "type": "string",
        "format": "uuid"
      },
      "Workspace": {
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "slug",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/WorkspaceId"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "WorkspaceDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Workspace"
          },
          {
            "type": "object",
            "required": [
              "locations"
            ],
            "properties": {
              "locations": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WorkspaceLocation"
                }
              }
            }
          }
        ]
      },
      "WorkspaceId": {
        "type": "string",
        "format": "uuid"
      },
      "WorkspaceLocation": {
        "type": "object",
        "description": "A workspace location as presented to the UI (node join included).",
        "required": [
          "node_id",
          "node_name",
          "node_status",
          "path",
          "dirty"
        ],
        "properties": {
          "dirty": {
            "type": "boolean"
          },
          "git_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "node_id": {
            "$ref": "#/components/schemas/NodeId"
          },
          "node_name": {
            "type": "string"
          },
          "node_status": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "worktree": {
            "type": "boolean",
            "description": "This checkout is a linked git worktree of the workspace's primary repo."
          }
        }
      },
...
    }
  }
}